requirements.txt
dash
gunicorn
pandas
numpy
requests
ccxt
plotly
scikit-learn
psycopg2-binary
bcrypt
